+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 * Add files_changed
+* Tables for display
 
 
 0.1.2 (2017-07-05)