+++ HISTORY.rst
@@ -2,7 +2,7 @@
 History
 =======
 
-0.1.3 (unreleased)
+0.2.0 (2017-07-05)
 ------------------
 
 * Add files_changed